no stable version, root not working, CurrentUser unfortunately null

+++ Model/Entities/Activity.cs
@@ -21,7 +21,7 @@ public class Activity
 
     [Required]
     [Column("WEIGHT")]
-    public decimal Weight { get; set; }
+    public float Weight { get; set; }
     
     [Required]
     [Column("REPETITION")]
